FIX THE SERVER
There was previously massive lag when connecting to the server from
the web client. It turns out that this was due to the line in question
in this commit, which is called whenever the browser connects to the
the server via socketio. This line was fetching entire chat objects
including the `messages` field, which contains thousands of messages
for some chats, so a very large array was being loaded into memory.
Selecting only the necessary field, `_id`, avoids this issue.